move the server starting code to separate file
making some big organization changes. May have broken something
with printing `buffer.` Maybe it was already broken? Band-aid
fix in place for now, wanted to upgrade it later anyways.
