+++ Makefile
@@ -2,9 +2,9 @@ CC = clang
 CFLAGS = -Weverything -std=c99
 BIN = controller
 
-$(BIN): main.o
-	$(CC) $< -o $@ $(CFLAGS)
-%.o: %.c
+$(BIN): main.o mcserver.o
+	$(CC) $^ -o $@
+%.o: %.[ch]
 	$(CC) -c $< -o $@ $(CFLAGS)
 format:
 	clang-format -i -style=file *.[ch]
